simulated-gps-producer/pyproject.toml
[project]
name = "simulated-gps-producer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiokafka>=0.12.0",
    "apache-airflow==3.0.3",
    "apache-sedona>=1.7.2",
    "attrs>=25.3.0",
    "geopandas>=1.1.1",
    "pandas>=2.3.1",
    "pyspark==3.5.6",
    "setuptools>=80.9.0",
    "shapely>=2.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
